requirements.txt
django==4.2.3
load-dotenv
pandas
seaborn
matplotlib
requests
mssql-django
pyodbc
django-mathfilters